backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.1
requests==2.31.0
soundfile==0.12.1
python-dotenv==1.0.0
Pillow==10.0.1
ffmpeg-python==0.2.0
torch>=2.1.0
numpy==1.22.0
opencv-python==4.7.0.72
librosa>=0.10.0,<0.11.0
TTS>=0.21.0,<0.23.0
